requirements.txt
apex
tqdm
scipy
numpy
imageio
pillow
torch>=1.0
torchvision
tensorboard
tensorboardX